Rename ts stdlib file to be consistent
