+++ apps/stdlib/src/websocket-interface.ts
